Add files via upload
This Power BI dashboard was inspired by my experience with the Dataquest Fifa tutorial. This is my original design, an interactive dashboard showcasing the infamous "Pace Abuser" strategy on Fifa 22. Players from Europe's top 5 leagues with pace greather than or equal to 85, finishing greater than or equal to 85, and sprint speed greater than or equal to 80 were compared in tabular and cartesian manners. An interactive slicer lies on the far right, along with a count of how many of these players are in each league on the bottom. https://www.youtube.com/watch?v=ukwBAi4ytms