Add files via upload
The 2022 soccer/football season brought us many qualified nominees for this year's "The Best" honor. Balon d'Or winner Benzema and his French compatriot Mbappe had historic seasons for Real Madrid and PSG, respectively. Furthermore, Mbappe's hat trick in France's unsuccessful bid to retain the World Cup (WC) will go down as one of the greatest WC final performances of all time. In the WC's storied history, only Geoff Hurst managed to score a hat trick in the tournament's championship match (1966, England). However, the favorite for this prize was none other than the Argentine magician, Lionel Messi. His tenure at the top of world football is unparalleled and was accentuated by his rivalry with the enigmatic Cristiano Ronaldo. Messi's slow start to life at PSG decimated his chances at the Balon D'or, but a strong start to his second season in Paris and a breathtaking WC run made him a shoe in for Fifa's award, which is voted on by each nation's coach, their captain, and a media representative. Fifa released the voting, which is visualized in this Power BI dashboard. An interactive map allows you to see the nationalities of each voter, along with slicers and various charts that emphasize just how dominant the voting for Messi was.